apps/media: Add 'pts rate' to packet definitions.
Packets, as defined in the transport mojom and in the framework's
Packet class, have a presentation timestamp whose units can vary.
So far, we've been relying on conventions: audio pts's are always
in frames and video pts's are always in nanoseconds. This isn't a
great state of affairs: the intention has always been to express pts
units explicitly in some way.

The design decision reflected in this CL is to have each packet
express the pts units individually. An alternative approach is to
declare the units in the media type, but this has a few practical
issues (like the fact that many components that look at packets don't
know the media type). So far, media type has only been required for
understanding packet payloads. Packet metadata has been media type
agnostic, and it would be good to keep things that way.

Packets now include rate information in the form of a rational value,
specifically the ratio of two 32-bit unsigned integers. In the mojom,
these integers are separate (pts_rate_ticks and pts_rate_seconds). In
the framework Packet class, we use a TimelineRate (pts_rate()).

A mojom packet struct's pts is expressed in seconds by the formula:

  packet->pts * packet->pts_rate_seconds / packet->pts_rate_ticks

or, for framework Packets:

  packet->pts / packet->pts_rate()

The pts rate of a packet can be altered without changing the packet's
effective pts so long as these formulae produce the same value before
and after the adjustment.

The framework Packet implementation has two new methods relating to
pts. GetPts(TimelineRate) returns the packet's pts in the specified
units. SetPtsRate(TimelineRate) adjusts the packet's rate without
changing the packet's effective pts.

The common packet consumer implementation (MediaPacketConsumerBase)
will now adjust the pts rate for all incoming packets if asked to do
so. The SetPtsRate method sets the desired pts rate to be applied to
all packets.

Change-Id: I3418444ab187baacefd883a51d414cdda9613f5c