[docs][syscalls] Update futex syscall docs
This makes two changes. First, it changes the declarations of these to
use the mx_futex_t type (this type exists since we currently can't name
an atomic integer the same way between userspace C code and kernelspace
C++ code).

Second, it changes the documentation of the possible return values of
wait and wake. These were changed en masse with many other functions,
but in this case the change is inappropriate. More importantly, it also
does not match what the functions actually return. Returning
ERR_ALREADY_BOUND ("the named entity is already owned or controlled by
another entity") implies that futexes necessarily exist in an owned
state that is correlated to the value located at that address. But they
don't actually have such an ownership semantics, and the kernel only
needs to be able to tell you if you are racing during the call. As such
ERR_BAD_STATE ("the system is unable to perform the operation in its
current state"), which the functions in fact already return, makes more
sense.

Change-Id: I2d5aba79217499bb9c64469d01fbfd101fd02811