[crashlogger] Catch crashlogger crashes
This installs an exception handler for the main thread
and then watches for exceptions from it on a separate thread.
This is useful for debugging as we now get crash dumps of the
crashlogger.

This is also important in order to report to the kernel
that the original crashing thread has been processed.
Otherwise anyone waiting on it will wait forever.

Change-Id: Ie51e2663aabe18993e92108d657b5e32e35f4658