[devmgr] Refuse to install remote handle if vnode already bound
The old behavior presents some problems in the context of properly
unmounting filesystems. If a filesystem is busy, but vn->remote
becomes closed, then the filesystem may not be correctly unmounted.
Instead, it will simply be inaccessible from the root filesystem.

A more sensible behavior involves refusing to mount the remote handle
if a vnode is already bound. To re-open the vnode for installation
access, the remote filesystem / socket should be removed by a distinct
unmount / removal process, which will also help unify the process of
removing a filesystem binding from a vnode.

Change-Id: I717a0535bdc2ce6a1e63fcf867bc4be10df44a9e