[kernel][magenta] Basic skeleton of Job dispatcher
The expected code. A job object is-a dispatcher and
contains a list of childs which are either job dispatchers
or process dispatchers and contains a raw pointer to
the parent job.

There is a root job, which is held by magenta.cpp and
that won't be exposed directly to user-mode.

No syscall api yet.

Change-Id: Ib484e3771e42222f7f43876279596f045b801ec4