[kernel] Precompute the kernel stack top in arch_thread_initialize()
Rather than computing the stack top on every context switch on x86, we
move the calculation to arch_thread_initialize() and save the value in
thread_t.

This saves a small amount of computation.  This will help with adding
an acpi_set_interrupts_enabled() syscall where the syscall will need
to get at the stack in order to modify the flags that the syscall sets
on return.

This does not save any computation on ARM32/64, but we make the same
change there for consistency.

Change-Id: If140a08951b07ed7babd0b01a98fcb0db6be3fdf