[FAT] Fix bug related to invalid '..' cluster.
The FAT specification requires that the cluster associated with a '..'
entry ALWAYS be zero if '..' actually refers to a root directory,
regardless of the FAT type (FAT16 or FAT32). Previously, I was just
returning whatever value for "StartCluster" I had chosen for FAT roots.

Previously, the mechanism by which ThinFS opened directories was through
the Dcache, using the starting cluster as a key. However, although a
"start cluster" of a directory is a unique key for a directory (like and
inode number in a more typical Unix-style filesystem), it is not the
ideal key for a root directory.

Instead, I have updated the Dcache to use an "ID" field of a directory
node as a key. This ID field will *typically* be a start cluster, but it
is guaranteed to be zero for root directories. This way, accessing a
root directory by reading a '..' direntry will not require additional
information about the FAT type, nor will it require reading the actual
root cluster from the BPB.

Change-Id: I7da7d9bd9afb5fd1291227d41f1fb1c5c2e35653