[utils] Allow adding overlapping regions in region-alloc
When adding regions to the set of available regions in the
region-alloc utility library, add a flag which permits the user to
specify that it is OK for the regions being added to overlap with the
existing available regions.  If the flag is set, the set of regions
available for allocation will become the union of the existing regions
and the region being added.

It is still not OK to attempt to add a region which intersets any
currently allocated region.

Change-Id: I64d19bbea7197a6ef38ed92cb3743e4392d35812