Move I/O to a separate thread
epoll/select is currently not available and there is no syscall to get
a handle from a socket fd that can be waited on for events (yet).
Since we are stuck with blocking reads and writes for now, this CL
moves the I/O logic to a new class called IOLoop that manages threads
dedicated to read and write operations.

Change-Id: Ica054b7919f0e7a86fd3882c0c29cf3df724d18d