[syscalls] mx_debug_task_get_child() -> mx_object_get_child()
Make this API a bit more generic in preparation for its use
with Resource Objects.

Take a rights mask similar to mx_handle_duplicate() and require
that the rights be a strict subset of the rights of the parent
handle, or the magic value MX_RIGHT_SAME_RIGHTS.

Shuffle all sys_object_*() syscalls into syscalls_object.cpp
while we're in the neighborhood.

Change-Id: I69322ac59f527a05a0c32f8027b8558e708e49fc