[udev][usb-xhci] Fix problem enumerating certain USB 3.0 hubs when attached at boot
In some cases the hub is already showing up as connected (bit CCS is set in PORTSC)
but the connect status change bit (CSC) is not.
To work around this problem, we now now act as if CSC is set if we see CCS set
when we first boot.

Also, now in xhci_start_root_hubs() we reset the root hub ports rather than power cycling.
Power cycling was causing us to lose the CCS bit, resulting in us losing track
of the fact that the hub was connected.

Change-Id: I2c18c721a839432d53b25b7c83d15c782dba5981