[mozart] Refactor input
Refactor Mozart input system by splitting HID interpretation and Mozart
Input Event generation.

The InputReader is now in charge of monitoring /dev/class/input for
devices and monitor event handles associated with file descriptors
for entries in the directory.

An InputDevice uses the HID descriptors to fill InputDescriptor which
are only used for now to get range values for touchscreen and stylus.

Once data is available and InputDevice will read and convert HID
reports into InputReports which are forwarded to the InputInterpreter.

The InputInterpreter converts the InputReport into Mozart::InputEvent
using the InputDescriptor if necessary, then forwards the InputEvent to
listeners.

Change-Id: I148f2b9c9ee330467765993d991b8a2b4cbe87b1