[devmgr] Fix rename bug when dnodes of varying lengths are used
When a dnode is renamed, the underlying structure can (generally
speaking) not be re-used: dnode structures contain a final element,
"char name[]", which is of variable length. Previously, the rename and
"dn_move_child" code assumed that "char name[]" was a structure of
length DN_NAME_MAX, so renaming from a shorter to a longer name would
cause an out-of-bounds memory access.

This patch also adjusts the organization of "memfs_rename".
It strives to:
  1) Make the distinction between validation and modification more
  clear, and
  2) Re-use existing dnode modification functionality, rather than
  relying on new codepaths.

Change-Id: I153c63f5636fc7f9c1787b085e60ccac4c74d4ec