apps/media: replace TimelineRate::Nano with TimelineRate::NsPerSecond
TimelineRate::Nano was defined to express a pts rate in nanoseconds,
specifically a billion ticks per second. It's value was set to
1/1000000000, which is inverse of what's needed for this purpose.
This CL corrects the value (to 1000000000/1). The name of the
constant has been changed (it's really more 'giga' than 'nano') to
make more sense in the context in which it's used.

Change-Id: Ia5efcd1166e8703b2426e265d67321aeb956ff2f