[musl][tsd] Initialize main thread tsd in exactly one place
Upstream this is done lazily depending on subsequent thread creation and
tls access. Since we do not specialize the single threaded case, do this
in one place (in particular, not at new thread creation time, and not in
different places in the static or dynamic link cases (i.e. __init_tls)).

Change-Id: Id60b124691a04181b43991a834f8a6dd21f3228e