Make the build-bootable-usb-depthcharge script install depthcharge.
Now that depthcharge is buildable in the Fuchsia environment, make the
aforementioned script build it and install it to the EFI system partition. The
script will default to the "dev" version of depthcharge for now, which will
be able to either netboot or boot the image from the USB key or system disk in
the future. For now it boots from the system or from the USB key.

Other versions will support only netbooting, or only booting through verified
boot. Once those are ready to use, the script can be extended to allow picking
which version you want.

Change-Id: I094a9062ed272487ed9268da0e021cd71642cff6