[rpi3][usb][dwc] Raspberry PI 3 USB Root Hub
This commit adds a fake root hub to the low-level Raspberry PI USB driver.

Since the root hub host port is onboard the USB controller, we intercept any
requests destined for the root hub in the low level driver and handle them
without scheduling them on the bus.

To the driver, this hub appears as an ordinary hub.

Change-Id: I665f5c636dd362de7866b88878c639fd410274a9