Get smarter about scheduling updates.
Simplified the interface to compositor backends so they are only
responsible for providing timing information for individually
scheduled frames.  Calculating update and snapshot times is now
done in a separate Scheduler class which is independent of the
backend.  This approach makes fewer assumptions than the old
Vsync scheduler did.

Moved rasterization to a separate thread.

Improved the Scheduling.mojom to clarify the purpose of the various
timing parameters which are exposed.  Also moved towards emphasizing
presentation_time as the ideal timing reference for synchronization
and animations.

Switched to using ftl::TimePoint and friends everywhere.

Change-Id: Ifa7ba377437546898b579883432468f129d7342f