[mx] Add libmx.a
This library contains C++ wrappers for the syscalls that help avoid type
errors and handle leaks. The library aims to inject as little opinion as
possible on top of the syscalls and does not depend on STL or libc.

The library builds with both the Magenta and Fuchsia build systems and
is not part of the sysroot.

Change-Id: Ie771b594eaf7f79da215abd0135c622e7821024f