apps/media: add video player example
This CL adds a video player example app. It uses a file reader for
the time being. When the http service is ready, I'll add the ability
to read from a URL. The player app uses the in-proc video renderer
from apps/media/cpp. The next step is to create a new video renderer
that's a mozart ViewProvider. The video player will just embed that.

Change-Id: I6a4af03220d3f8e0fcc168220d00a3c6d4c41592