[udev] Add block alignment filter driver
Many POSIX programs operating on device drivers expect that the
underlying device will be accessible like a regular file: requests can
be made in requests that are not aligned to the block size, of sizes
that are not a multiple of the block size.

This filter driver adds support for these operations, completing the
request rather than returning ERR_INVALID_ARGS. Though the non-aligned
case adds a small amount of overhead, requests that are already aligned
will incur minimal overhead -- this driver will act as a pass-through.

Change-Id: I8b411f3b7534fdc191d666dba05c30a4b6b0bf6e