Make fidl.gni work.
This brings in some config and build files that were local to mojo; depending on
how externally-consumable we want to make libfidl, we might want to keep this
local, or otherwise move it to //build.

There's also lots of mojo stuff we need to remove now that we don't use
prebuilts (and compile our generators/compilers).

This patch will let you use the fidl() template to generate fidl files and try
to compile them (which will fail until the generators are fixed to point to
libfidl).

Change-Id: I3ca2ad41ba2ea37443ec7c6f3bbce7b474f32fb6