apps/media: minor fixes to flog
This CL contains a few minor fixes to flog. Most importantly, there
was an issue when large numbers (1024) entries were requested by
the viewer using GetEntries. This is under investigation, and there's
currently no bug tracking this issue. Reducing the request to 64
entries is an effective workaround. The remainder of the changes have
to do with output formatting and the use of std::cerr where FTL_DLOG
should have been used.

Change-Id: I3ca75b471c97e2e779f13a24325027d2a8e8d945