[fat] Add support for flipping and checking multiple FAT bits
As mentioned in a comment in the code, Linux and BSD use distinct bits
to signify that the filesystem was not unmounted properly. To avoid
mounting a corrupted filesystem, this patch updates the FAT driver to
check and set both bits. This avoids both corruption of FAT filesystems
coming to Fuchsia for the first time and corruption of FAT filesystems
leaving Fuchsia and being used on other operating systems.

A corresponding patch has been sent to our fsck, which is upgrades it to
identify and fix both bits as well.

Change-Id: Iba39c648639d64ffb048a074bf9ca7f31dec669a