[fat] Updated starting cluster for empty files
The FAT specification states that any cluster value at or above a
certain threshold should be interpreted as "EOF" (for FAT32, this value
is 0x0FFFFFF8). Thus, a file with some clusters will contain:

  [Directory Entry]
  cluster 1         --> cluster 2 --> ... -> EOF cluster

Previously, for empty files, ThinFS would store the EOF cluster directly
in a directory entry. However, the FAT spec states that "a zero-length
file... has a first cluster number of 0 placed in its directory entry".

This patch updates the treatment of empty files to match this
specification.

Change-Id: I3562882343a60c0e54152f9b3dd7c8442dda75ce