[build] whitelist libraries for sysroot
The sysroot should be carefully curated, not blindly include
every library that we happen to build within the magenta build.

MODULE_EXPORT is now used to indicate a library belongs in sysroot,
either as static "a", shared "so", or both "a so".

MODULE_NAME can override the static library base name if desired,
similar to how it overrides the installed binary name for apps.

Change-Id: Id97163a46e6ade155e82ad88ebcb68d48e243f57