[appmgr] Support application launcher channel.
This change enables application launching from mxsh using the mechanism
formerly added for the old application manager.  On reflection, it might
be better to make the shell aware of its environment (or perhaps embed
it within an application that is) and simply use the application
launcher interfaces more directly but this is good enough for today.

Change-Id: Iccad56543bb5eb1c2201dcd8a3588d197c2964ac