[kernel][syscalls] Add object_signal_peer() syscall
My previous CL (14136) introduced user signals for all objects that are
waitable. This CL adds a new syscall so for two sided objects
the side that gets signal is clear

- object_signal(handle, clear_bits, set_bits)  signals the side
attached to the handle
- object_signal_peer(handle, clear_bits, set_bits) signals the
opposite side to the handle

If and object does not have 'other side' the syscall return
is ERR_NOT_SUPPORTED.

This CL does not address the Channel Object which needs
other changes first.

Change-Id: I83a8b0dff3027aa59f64933ba708ac9bccb67d8b