[kernel] Make wait_queue_destroy() stricter
There are currently four uses of wait_queue_destroy(): cond.c,
event.c, mutex.c and semaphore.c.

In all those cases, it indicates a race condition if any threads are
on the wait queue when it is destroyed -- see comment for explanation.

wait_queue_destroy() was already checking for this problem in debug
builds.  Let's make that the behaviour in all builds.

This makes the "reschedule" argument unused, so we remove it.

MG-354

Change-Id: I75819eaeb754dde7efb667906dc0a07b73bfddb9