[kernel] Remove unused function mutex_acquire_timeout()
This will simplify the task of making kernel/kernel/mutex.c faster in
the uncontended case, because it removes the need to handle timeouts.

If we need this operation in future, we can add it back.  However,
claiming a mutex with a timeout isn't a very commonly-used operation,
so it's fairly likely that we'll never need it.

MG-362

Change-Id: Iba4b76e5bb52e07cf9837f1205187a962472c4cd