Rework FIDL Dart bindings
The new Dart bindings work more like the C++ bindings. Instead of
inheriting from a Stub object, implementations of interfaces compose in
a Binding object.

This CL also introduces InterfaceRequest and InterfaceHandle classes that
play similar roles to those classes in C++. This removes the awkward
"unbound" constructor for Proxy and Stub objects.

This patch also cleans up the public interface of Proxy and Binding
objects. Proxy objects now grab only a single name (ctrl) and Bindings
objects grab zero names. Also, the ChannelReader interface is hidden
deeper as an implementation detail, similar to Router and Connector in
the C++ bindings.

Finally, this CL renames a number of classes for clarity.

Change-Id: I7df074fca65368ec2aa596b71c75172d4fa00ab6