Beginning support for Rust bindings
This commit is work in progress. It contains the beginnings of a
library crate (fidl), containing serialization of most fidl data types
(primitives, arrays, maps, structs, unions, interfaces, interface
requests, and nullable variants where appropriate).

In addition, it begins to adapt the existing rustgen to support the
new style serialization. A simple (hello) protocol works, but more
complex interfaces have not been tested.

Other work in progress: versioning of structs is not handled well.
The service name metadata is encoded, but there is no API for
accessing it. Some performance is left on the table (particularly
Vec<u8> and friends).

There is a placeholder runtime that is very thread-heavy. The idea
is to transition to a much more asynchronous runloop, based on
futures-rs, with fairly minimal disruption to types. But currently
thread usage will be bad as services scale.

Change-Id: Ib4be745f642210ee0c72a596efa5bd10686d6a24