[bootloader] Fix printf warnings under Clang
This cleans up the printf formats in bootloader/ to use the
PRIx64 macro from <inttypes.h> rather than presuming that
uint64_t is unsigned long.  (In Clang for the Windows target used
for the booloader, it's unsigned long long instead.)

It also fixes the makefiles so that 'make gigaboot' works.

Change-Id: I98de7c4c9abe429d101a9cc7f0c06767e8a72bfd