[kernel][ulib][utest] Use {} for all-zeros initializers
An empty initializer list {} is sufficient to guarantee an
initialization to all zeros (pedantically, an initialization
equivalent to the default initialization of a static with no
explicit initializer).  An initializer like {0} for a struct
containing multiple fields generates a warning from Clang
(-Wmissing-field-initializers, on by default).  This is a useful
warning in general, since it helps find accidental omissions.
There is no warning for just {}, so use that everywhere that "all
zeros" is the intent.

Change-Id: I9d67bfe15723fc1c53422b5529a433385a4ff274