[ulib][region-alloc] Use incomplete type, rather than empty structs, for opaque pointers
An empty struct has different semantics in C and C++, so Clang
warns about them (-Wextern-c-compat).

The traditional method of declaring a type for use as a distinguished
opaque pointer type is to use an incomplete struct type.  Pointers to
incomplete types are always fine.  This is better than an empty struct,
because it's completely impossible for any user to define an object of an
incomplete type.

Change-Id: I1790d921ee1a0c4971fde1310eebf38628d90470