[bitmap] Adding num_elems_ field to RleBitmap
This CL adds a field to track the length of the elems_ list in
RleBitmap.  The upcoming HybridBitmap needs to know how long the list is
in order to know when to switch to a RawBitmap.  The alternative to
tracking it here is to use the O(n) approach of
DoublyLinkedList::size_slow().

Change-Id: I1cee9e975347cae35fd5708575f1b97b73ad1441