[bitmap] Add RawBitmap::Reset
This CL adds a method to reset the RawBitmap, possibly to a new size.
Besides being useful in and of itself, this will allow the
HybridBitmap to avoid lots of new/deletes when changing between Raw
and Rle; instead it can simply set Raw's size to zero or clear Rle's
elems_.

It also does a minor refactor to RawBitmap, and tries to consolidate
all the bit indexing logic into the functions in the anonymous
namespace.

Change-Id: I0ac01a2b3ded86cf10614aeaf96e854492a35364