Less opinion (handles and message buf allocation), more docs
Change the mechanism for transferring handles from a message buf from
an iterator to a take_handle() call, the latter of which is much
closer to what fidl deserialization needs.

Add a Channel::read_raw() so that it doesn't do any reallocation by
itself.

Add lots of docs, most objects and methods are now covered.

Change-Id: I853a376dcaeecfadfdc91e5a9913f55c96584d6f