Add Bootstrap program to help set up environment.
Bootstrap will generally be one of the very first applications
which is run.  It takes care of setting up access to global
system services then launches another application within the
environment which was created.

Most of the code was simply moved out the Mozart Launcher to
Bootstrap, but we now offer a few other fancy features such as
a configuration file and the ability to register additional
services on the command-line.

It's quite useful to represent Bootstrap as a separate application
as it enables testing of lower level system services in isolation
from those which bear more policy (such as 'device_runner').

Bootstrap is designed to be fairly robust.  If any of the services
dies, they will be restarted automatically the next time an
application attempts to connect to that service.

Currently Bootstrap is responsible for configuring the view manager.
This will be removed in a future patch.

See README.md for more details.

Change-Id: I4e54cda117822c93acd47862d89370da2983a99d