[udev][gpt] Returning GUID as uint8_t array, not string
This change relocates the "uint8 to string" utility to the GPT library,
and adds a dependency on the GPT library where previously applications
(devmgr and lsblk) sent ioctls directly to the GPT driver.

This change intentionally copies the "uint8 to string" utility from the
gpt driver to the gpt library -- the copy is not shared, to avoid making
a dependency from a device driver on a larger ulib library.

Change-Id: I6368a8409d5322c1424c2ae1c550744f1b073fff