[utest][debugger] Clean up sloppy asm
The aarch64 instruction used to materialize a PC value was bogus.
(It's an assembler bug that it was accepted at all.)  This went
unnoticed because that value was unused except for printing it out.
The x86_64 instruction used to materialize a PC value was using
unnecessary indirection and a complex reloc type rather than the
simple thing the compiler would have done.

Clean up the assembly code for both aarch64 and x86_64 so it uses
the canonical method to materialize a PC.  Add a runtime test that
the PC value so computed matches the fault PC as intended.

Change-Id: Iacb3455c6af970ddd547fcb504b20fe9042595b8