[kernel] Fix race between AddThread() and Kill()/Exit()
AddThread() checks state_ and refuses to add a thread if the process
is DYING or DEAD.  However, before this change, it doesn't claim
state_lock_, so that check is ineffective: it can race with
terminating the process (via SetState()), which sets state_.  As a
result, a thread can get added after process termination has been
started.  If that happens, it's possible that process termination will
never complete.

The obvious fix is to claim state_lock_ in AddThread().  That doesn't
work by itself, because there are two routes through which AddThread()
is called:

 1) Starting a process's initial thread:
      sys_process_start()
      -> ProcessDispatcher::Start()
      -> UserThread::Start()
      -> ProcessDispatcher::AddThread()

    ProcessDispatcher::Start() claims state_lock_, so AddThread()
    can't claim it too.

 2) Starting subsequent threads:
      sys_thread_start()
      -> UserThread::Start()
      -> ProcessDispatcher::AddThread()

This change resolves that by removing ProcessDispatcher::Start() and
moving its functionality into AddThread().  This makes
sys_process_start() and sys_thread_start() more uniform because both
now call UserThread::Start().

MG-355

Change-Id: I2ad5a61d1c56268171e1389dd23126f5c62aa8b3