[kernel] Simplify locking in UserThread::Kill()
Accessing state_ concurrently, outside of a lock, without doing it as
an atomic access, is somewhat dodgy.  It's a race, so it's technically
undefined behaviour.  It would be better not to do this.

It turns out we can remove this by changing Exiting().
UserThread::Exiting() is called on the corresponding thread, and it is
only called once per thread, so its UpdateState() and RemoveThread()
calls don't need to be done while holding state_lock_.  This removes
the problem where RemoveThread() was called while holding UserThread's
state_lock_.

Change-Id: I95d8827f78378eec7218f1375634f5a80498a0fa