cloud_provider: implement object API (using Firebase).
This patch implements the object cloud storage API. For now, this just
uses Firebase, which we already need and use for the commit
notifications, as a stop-gap to allow experimentation.

Using Firebase has the limitation of not being able to store objects
bigger than 10MB. We are going to switch to use GCS instead when it
becomes neccessary.

LE-62 #done

Change-Id: I2465cb4ee198f8e2e67daf69c74c2e656370afb7