[devhost][minfs] Add support for MXRIO_SYNC to devhost
If devhost is able to receive a "MXRIO_SYNC" command, then usage of
"underlying storage" becomes much more flexible for our filesystems.
In the current implementation, filesystems are REQUIRED to be mounted on
storage which implements the device protocol. With this patch, however,
filesystems can be mounted on storage other than physical devices.

Additionally, this patch allows the "sync" and "fsync" commands in
unistd to be applicable to both files and underlying storage.

Change-Id: Ib6143bfb8f336f3aa91dbc122d143682aaf9abf8