cloud_sync: Add PageSync (upload only).
This patch adds an elementary version of PageSync, based on the
previously added CommitUpload. PageSync in this form:

 - uploads locally created commits and objects using CloudProvider
 - retries on error using a configurable backoff policy

In the next CL we will add the other direction (download).

Once Network Service or some other entity is able to deliver connection
status notifications, we will use it as a signal to kick off retries.

LE-42

Change-Id: I507391af7322459c3c84e571af7f55e7b52d4a21