[mutex][utest] Fix timing issue in c11 mutex test
Currently the kernel can truncate timeouts by up to a millisecond,
more or less, as this is the quantization of the lk_time_t timers. So
fudge the test to only need a timeout within a millisecond as long as
requested.

cpu's nuc bots are currently experiencing this, and I could also
reproduce the test failure occasionally on qemu.

Change-Id: Idcc585f015666c459af561bf68fd5d847165564c