Add more flutter module examples
This CL contains two more flutter example modules that are in parent-
child relationship. The parent spawns the child module using the Story
interface, and then composes the child module into the UI using the
flutter ChildView.

Both modules have increase / decrease buttons, which can be used to
modify the counter value. The new counter value is written to a shared
Link object. Both modules also implement LinkWatcher interface, and
they get notified with the new changes and update the UIs in turn.

Change-Id: Ifa2c0ea79bf11950cbb3504597c2620f7f1ee513