cloud_sync: implement cloud-to-local sync.
Also known as "download", this patch extends PageSync to also handle
pulling down commits from the cloud and storing them in storage. The
backlog of remote commits is downloaded first, then a cloud watcher is
set to track new remote commits appearing in the cloud provider. Remote
commits are added to storage in the order in which they were added to
the cloud provided.

In order to track which remote commits were already fetched, we keep
track of the server-side timestamp of the last commit we added to
storage. As this information needs to be persisted through reboots, we
store the timestamp itself in storage using a dedicated API
(Get/SetSyncMetadata()).

There's a few follow-up changes to do after this:

 - LE-76 allow to distinguish connection errors from data errors in
   cloud watcher, and retry setting the watcher on connection errors
 - LE-78 make storage AddCommitFromSync() api asynchronous
 - LE-79 teach storage to accept synced commits in batches

LE-42 #done

Change-Id: I976044770cbafefa0a109e72829fc07fc56ca424