thinio: Prevent infinite loop in flush
Rather than silently looping forever on failure (current behavior), or
returning an error from flush to a different function (the result of
recording a flush error), simply silence the error from Flush.

The entire error handling scheme in thinio needs to be re-worked;
propagating errors to unrelated functions is NOT acceptable beahavior
for a block cache.

In the meanttime, this prevents a bug in Magenta for any block devices
which do not implement IOCTL_DEVICE_SYNC.

Change-Id: Iee150d9b414261e98c96054ef41c2e9c406482a0