Add /system/apps to the application path.
As part of this change, we split the configuration file so that
applications.config includes startup.config which is defined
separately (and is optional).  Meanwhile applications.config
now contains /system/apps as a search path for applications.

Loading applications is very slightly factored out to a new
ApplicationLoader class.  In the future we will want to give
application hosts the ability to provide their own loaders.

This change enables a much more compact way to run apps, eg.

  > @ bootstrap launch noodles_view

Change-Id: I02f74a6f697599987cf0505a71510564d37a7fb4