[kernel][magenta] Jobs reformation
It all started rather innocently.. but the changes needed to happen

1- Before, the job had refptrs to the child jobs and child processes and
and each process a refptr to the parent job. That is a ref cycle! Now
following the proces-thread model, the job has weak pointers to the
processes and the process has a refptr to the parent job.

2- Before, there was a JobNode class that was polymorphic on either job
or process and the Node was LinkedListable. The job had a single list of
children. Now all that is gone; both the child processes and child jobs
are directly listable in two separated lists in the job. Besides removing
a lot of code we know don't need heap allocations.

3- We know correctly add and remove processes from the associated
job. This needed to be in this CL because of the weak pointer contract
as explained in #1

4- A new diagnostic command "`mx jb <pid>" so given a process koid
we can show the job its in along with the parent job and so on up to
the root along with sibling processes and child jobs.

Here for 1040
[00068.226] 01125.01153> process 1040 [bin/devmgr]
[00068.226] 01125.01153> in job [1029]-->[1024]
[00068.226] 01125.01153> no processes
[00068.226] 01125.01153> - 1053 child job (8 processes)
[00068.226] 01125.01153> - 1072 child job (6 processes)

Here for 1076
[00075.713] 01125.01153> process 1076 [crashlogger]
[00075.713] 01125.01153> in job [1072]-->[1029]-->[1024]
[00075.713] 01125.01153> no jobs
[00075.713] 01125.01153> - 1111 proc [netsvc]
[00075.713] 01125.01153> - 2089 proc [mxsh:console]
[00075.713] 01125.01153> - 2374 proc [mxsh:vc]
[00075.713] 01125.01153> - 2423 proc [mxsh:vc]
[00075.713] 01125.01153> - 2454 proc [mxsh:vc]

Change-Id: I6185b7148126d460e38f09fa3ce209bd28c007b9