[devmgr] Add support for 'unmount_node' in devmgr.
Previously, Magenta had no mechanism to unmount specific filesystems on
a mount path -- vfs-mount was responsible for unmounting all
filesystems, in the instance of powering off.

This patch adds support for a more focused unmount, in the form of
"ioctl_devmgr_unmount_node". Devmgr's vfs implementation of this ioctl
is similar to vfs_open, in that it may be forwarded to a remote
filesystem. However, there is additional handling in "vfs_unmount_node"
to ensure that a request to unmount "path" will target the hosting
vnode at "path", not the mounted, remote root vnode at "path".

Additionally, updated vfs-mount to handle errors more appropriately,
since we are no longer assuming that it will only be called when
powering off.

Change-Id: I28a02dfbd8616b69f98bc45583a7fc82716e7895