[bitmap] Use size_t throughout
This CL changes the interface to always use size_t to be more portable
between 32 and 64 bit platforms.  The only requirement on size_t's size
is in RawBitmap, where it must match unsigned int, unsigned long, or
unsigned long long in order to pick the correct GCC builtin for counting
trailing zeros.

Change-Id: I42acf1f9363586a10aac204efc0d6b21cb5f653a