[ulib][magenta] Work around dubious LLVM optimization in mx_status_get_string
The non-PIC-friendly code generated by LLVM for this function is the
last thing preventing the Clang build of Magenta from working.  We'd
like to fix LLVM to generate PIC-friendly code for this case.  But
in the meantime, work around the problem by tweaking the code so
that LLVM decides against performing the problematic optimization.

Change-Id: I62892359ced317f01850c9bf6d9004b73cf37c09