Refactor trace engine, writer, and associated tables.
Moved low-level portions of trace writer into TraceEngine together
with related state variables.

Added TraceWriter as a way to keep track of active writing operations
using RAII semantics.  This change removes the need to repeatedly
check whether the trace engine is running for each individual call
to methods such as RegisterString() and RegisterCurrentThread() which
happen frequently over the course of writing a single record.
This also lays the groundwork for implementing a proper shutdown
protocol to avoid race conditions (in a future change).

Added CategorizedTraceWriter as a way to avoid having to resolve the
category twice when writing out a record (once to determine if the
category is enabled and again to get the StringRef).

Implemented lock-free string and thread tables.  The string table
also keeps track of enabled categories to reduce the number of
lookups required to write a record.

Added a few more trace event macros and some basic tests to ensure
that they all expand correctly.

Change-Id: Ib0cd741c09ccbb35522dfa10e421f53c3e70c80f