[kernel][timer] fix bug where timer can be fired during or after a timer_cancel()
A hard to hit race where a cpu can attempt to cancel a timer that is already
in a processing state on another cpu could cause the timer to appear to fire
after the cancel has completed.

Fix by adding a two variable interlock in the timer structure so the canceller
and the processor can make sure they do not step on each other's toes.

Change-Id: I54648ab65582e2b382c8e62ead0a8eb1bdacd410