GN changes for Rust
This patch builds Rust crates on a per-GN-target basis. To do so, it
creates a map from each fidl file to the GN target that contains it.

Most crates in the fuchsia tree now build. Some tests fail, for a
variety of reasons, including limited support for fixed size arrays.

When GN dependencies are missing, the generated code has
::UNKNOWN_CRATE, which obviously will also fail to compile. This could
easily be changed to be a fatal error at generation time.

Change-Id: I89cc410a570277a6f101d0f20ff2510ba9f350f7