cloud_sync: add remote commits to storage asynchronously
This patch teaches cloud sync that when it adds a new remote commit to
storage, it needs to wait for an asynchronous confirmation before
proceeding to add another one.

This doesn't change anything on the storage side yet, but in some future
patch storage will be able to use this contract to fetch all associated
objects before confirming, thus ensuring that it has fully constructed
one commit before receiving its children.

LE-78 #done

Change-Id: I5f9b051942c4be269712e2e659e2830620f3038f