cloud_sync: add PageSync OnBacklogDownloaded callback.
Adds a callback that will be called at most once after Start(), when all
remote commits added to the cloud between the last sync and starting the
current sync are added to storage. This can be used by the client to
delay exposing the local page until it catches up with the cloud.

LE-94 #done

Change-Id: Ifcc65fcbeee1ad0c0607258df84628ab0efe015d