Replace Dart HandleWatcher with HandleWaiter
The HandleWatcher design was broken because there was no way of canceling a
wait synchronously because the wait actually happened on a background thread.

The new HandleWaiter interface mimics the AsyncWaiter design that we use in C++
and waits on the current threads mtl::MessageLoop. This lets us synchronously
cancle waits.

In order to work properly, we can no longer leak bound interfaces. If there are
any outstanding waits when a HandleWaiter is garbage collected, it crashes the
application (except of the whole isolate is shutting down).

Change-Id: Ieac9667aa7c57b28875105c072fb6331f75c04ab