Use new trace provider protocol and fix races during shutdown.
The new protocol uses a fence rather than a callback to signal when
the trace has finished.  This neatly addresses problems such as
coping with the trace manager crashing and it provides a channel
over which we should be able to deliver signals related to
buffer filling and draining.

Unfortunately the usual ref-counted smart pointer types do not work
for us here as we have no way to atomically manipulate the ref-pointer
itself without grabbing a mutex.

So instead we make the reference counting behavior explicit and only
use a mutex to guard major state changes of the trace writer.

Change-Id: Ida5e54721125248902f55a52a54bc9b40854e6ba