[ddk][bug] Don't memset the stack in iotxn_alloc
When iotxn_alloc found an iotxn_t in the free_list it could reuse, it
tried to zero it, but used the address of the pointer to the transaction
rather than the pointer itself.  If priv->buffer_size is large enough,
this can blow away other parameters on the stack, including out which is
dereferenced shortly thereafter.

Change-Id: Id3c415cc390c91e203a310343cbc3daf79f60f11