[ddk][bug] Clear free flag on clone
When iotxn_clone finds an iotxn_t on the clone_list that it can reuse,
it attempts to clear the 'free' flag, but cleared the flag on the
source iotxn_priv_t rather than the target.  This leaves the flag set on
the clone, result in a 'double release' in iotxn_release.

Change-Id: I42b7ac99226bb69a3c99c63c341606cfc4b2d5c2