[minfs] Update 'len' during readdir
Previously, the 'len' field of readdir was unchanged, giving
'vfs_fill_dirent' the wrong impression that it always had a full 'len'
(typically, 8KB) of space to use when copying a dirent to an output
buffer. As a consequence, the output buffer overflowed for large
directories.

This patch fixes that issue, and includes an assertion to catch improper
input overflows more quickly.

US-99 #comment Fixes the bug

Change-Id: I9c6008f44192a30665b3ded8e4f84036cd3ea3f5