Add the todo example.
The application keeps a list of todo items in a Ledger page, randomly
adds and removes items, and prints the state after each modification.

The application takes a tag as an optional positional argument, so that
items being added by the application on different devices can be
distinguished.

To run the application:

```
@ bootstrap todo
```

or, to run with a tag:

```
@ bootstrap todo Alice
```

Change-Id: I9a992229e849bc9aa6a2eb8d5fac46817520931f