Add TraceHandler mechanism.
TraceHandlers allow apps  to participate more deeply in the
tracing lifecycle.  This is especially important for apps which
may need to perform expensive operations when starting and
stopping tracing, such as will be needed for the ktrace provider.

As part of this change, the writer's threading invariants had
to be changed somewhat to ensure that traces are always stopped
on the same thread as they are started which simplifies the
implementation somewhat.

Added a way to query the trace system's state and the set of
currently enabled trace categories.

Change-Id: I9039c414ba9bb42116ab40d40bfa3db96b404a83