Decouple TraceWriter from current thread.
When writing custom trace providers, we need to be able to ask
TraceWriter to output events about any thread with any category
or name (even those which might not be stored in string contants).

Factored out the thread and string registration over to the macros.

Moved the internal macros to a new file so that writer.h itself
does not need to rely on them.

Change-Id: Ie44127b8c1a49e6e2bea5d1578a75ecf6d3112dd